Unify authentication screens
This unifies both authentication flows (to II and to other dapps) by
showing the same authentication screen in both cases (with slight
variations). The "welcome" page is also simplified to lead new users
towards the correct flow, i.e. creating an anchor, while hiding
everything else (unnecessary stuff for new users like anchor recovery,
etc) behind another button. The welcome screen is only shown to new
users.
