+++ src/frontend/src/flows/login/index.ts
@@ -1,205 +1,16 @@
 import { Connection } from "../../utils/iiConnection";
-import { Ref } from "lit-html/directives/ref.js";
-import { addRemoteDevice } from "../addDevice/welcomeView";
-import { LoginFlowSuccess, LoginFlowError, LoginData } from "./flowResult";
-import { displayError } from "../../components/displayError";
-import { footer } from "../../components/footer";
-import { getUserNumber } from "../../utils/userNumber";
-import { html, render, TemplateResult } from "lit-html";
-import { icLogo } from "../../components/icons";
-import { mkAnchorInput } from "../../components/anchorInput";
-import { parseUserNumber, setUserNumber } from "../../utils/userNumber";
-import { registerIfAllowed } from "../../utils/registerAllowedCheck";
-import { authenticate } from "../authenticate";
-import { unknownToString, unreachable } from "../../utils/utils";
-import { useRecovery } from "../recovery/useRecovery";
+import { LoginData } from "./flowResult";
+import {
+  authenticateBox,
+  AuthTemplates,
+} from "../../components/authenticateBox";
+import { html } from "lit-html";
 
-/** Data and callbacks used on the login page */
-type PageProps = {
-  submit: (res: bigint) => void;
-  addDevice: (userNumber: bigint | undefined) => void;
-  recover: () => void;
-  register: () => void;
-} & Returning;
-
-/** Type representing either a returning user (with saved anchor and function to clear
- * said anchor) or a first time user */
-export type Returning =
-  | { returning: true; clearCache: () => void; userNumber: bigint }
-  | { returning: false };
-
-// We retry logging in until we get a successful Identity Anchor connection pair
-// If we encounter an unexpected error we reload to be safe
-export const login = async (connection: Connection): Promise<LoginData> => {
-  try {
-    const userNumber = getUserNumber();
-    const clearCache = () => {
-      const message = `This will forget your anchor number but your credentials will still be stored. To clear your credentials please check your device's documentation on how to clear WebAuthn credentials.`;
-      const confirmed = confirm(message);
-      if (confirmed) {
-        window.localStorage.clear();
-        window.location.reload();
-      }
-    };
-
-    const returning: Returning =
-      userNumber !== undefined
-        ? { returning: true, clearCache, userNumber }
-        : { returning: false };
-    const x = await new Promise<LoginFlowSuccess | LoginFlowError>(
-      (resolve, reject) => {
-        loginPage({
-          submit: async (userNumber) => {
-            const loginData = await authenticate(connection, userNumber);
-            if (loginData.tag === "ok") {
-              setUserNumber(loginData.userNumber);
-            }
-            resolve(loginData);
-          },
-          addDevice: (userNumber) => addRemoteDevice(connection, userNumber),
-          recover: () => useRecovery(connection),
-          register: () =>
-            registerIfAllowed(connection)
-              .then((res) => {
-                if (res.tag === "canceled") {
-                  window.location.reload();
-                } else {
-                  resolve(res);
-                }
-              })
-              .catch(reject),
-          ...returning,
-        });
-      }
-    );
-
-    switch (x.tag) {
-      case "ok": {
-        return { userNumber: x.userNumber, connection: x.connection };
-      }
-      case "err": {
-        await displayError({ ...x, primaryButton: "Try again" });
-        return login(connection);
-      }
-      default:
-        unreachable(x);
-        break;
-    }
-  } catch (err: unknown) {
-    await displayError({
-      title: "Something went wrong",
-      message:
-        "An unexpected error occurred during authentication. Please try again",
-      detail: unknownToString(err, "Unknown error"),
-      primaryButton: "Try again",
-    });
-    window.location.reload();
-    return Promise.reject(err);
-  }
+/* Template for the authbox when authenticating to II */
+const tpls: AuthTemplates = {
+  message: html`<p class="t-lead">Authenticate to manage your anchor</p>`,
 };
 
-export const loginPage = (props: PageProps): void => {
-  const container = document.getElementById("pageContent") as HTMLElement;
-  const content = pageContent(props);
-  render(content.template, container);
-};
-
-const pageContent = (
-  props: PageProps
-): {
-  template: TemplateResult;
-  userNumberInput: Ref<HTMLInputElement>;
-} => {
-  const anchorInput = mkAnchorInput({
-    onSubmit: props.submit,
-    userNumber: props.returning ? props.userNumber : undefined,
-    focus: props.returning,
-    important: props.returning,
-  });
-
-  const addDeviceClick = () => {
-    const userNumberInput = anchorInput.userNumberInput.value;
-    const userNumber =
-      userNumberInput === undefined
-        ? undefined
-        : parseUserNumber(userNumberInput.value) ?? undefined;
-    props.addDevice(userNumber);
-  };
-
-  const signin = html`
-    <div class="${props.returning && "l-stack"}">
-      ${anchorInput.template}
-      <button
-        @click="${anchorInput.submit}"
-        type="button"
-        id="loginButton"
-        class="c-button ${props.returning ? "" : "c-button--secondary"}"
-      >
-        Authenticate
-      </button>
-      <p class="t-paragraph t-weak t-centered">
-        Using this device for the first time?
-        <button @click=${addDeviceClick} id="addNewDeviceButton" class="t-link">
-          Add this device</button
-        ><br />
-        Lost access?
-        <button @click=${props.recover} id="recoverButton" class="t-link">
-          Recover anchor
-        </button>
-      </p>
-    </div>
-  `;
-
-  const signup = html`
-    <div class="${!props.returning && "l-stack"}">
-      <div class="l-stack">
-        <button
-          type="button"
-          @click=${props.register}
-          id="registerButton"
-          class="c-button ${props.returning ? "c-button--secondary" : ""}"
-        >
-          Create an Anchor
-        </button>
-      </div>
-    </div>
-  `;
-
-  const dividerReturning = html`
-    <div class="l-divider l-divider--text" aria-label="Other Options">New?</div>
-  `;
-
-  const divider = html`<div
-    class="l-divider l-divider--text"
-    aria-label="Other Options"
-  >
-    Returning?
-  </div>`;
-
-  const template = html` <section
-      class="l-container c-card c-card--highlight"
-      aria-label="Authentication"
-    >
-      <div class="c-logo">${icLogo}</div>
-      <p class="t-paragraph t-centered">Unlock the full power of Web3</p>
-      <article class="l-stack">
-        ${props.returning ? signin : signup}
-        ${props.returning ? dividerReturning : divider}
-        ${props.returning ? signup : signin}
-      </article>
-
-      ${props.returning
-        ? html`
-            <div class="l-divider" aria-label="Other Options"></div>
-            <p class="t-paragraph t-weak t-centered">
-              <button @click=${props.clearCache} class="t-link">
-                Clear identity anchor from browser
-              </button>
-            </p>
-          `
-        : ""}
-    </section>
-    ${footer}`;
-
-  return { ...anchorInput, template };
-};
+/* the II authentication flow */
+export const auth = async (connection: Connection): Promise<LoginData> =>
+  authenticateBox(connection, tpls);
